social_network/requirements.txt
django==5.2.1
djangorestframework==3.14.0
pillow==11.2.1
psycopg2-binary==2.9.10
geopy==2.4.1
